+++ RecatchLegendaryFish/Framework/ModConfig.cs
@@ -1,3 +1,6 @@
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.Serialization;
+using cantorsdust.Common;
 using StardewModdingAPI.Utilities;
 
 namespace RecatchLegendaryFish.Framework;
@@ -8,6 +11,24 @@ internal class ModConfig
     /*********
     ** Accessors
     *********/
+    /// <summary>How often each legendary fish can be caught.</summary>
+    public CatchLimitType CatchLimit { get; set; } = CatchLimitType.Unlimited;
+
     /// <summary>A keybind which toggles whether the player can recatch fish.</summary>
     public KeybindList ToggleKey { get; set; } = new();
+
+
+    /*********
+    ** Private methods
+    *********/
+    /// <summary>The method called after the config file is deserialized.</summary>
+    /// <param name="context">The deserialization context.</param>
+    [OnDeserialized]
+    [SuppressMessage("ReSharper", "NullCoalescingConditionIsAlwaysNotNullAccordingToAPIContract", Justification = SuppressReasons.ValidatesNullability)]
+    [SuppressMessage("ReSharper", "UnusedMember.Local", Justification = SuppressReasons.UsedViaReflection)]
+    [SuppressMessage("ReSharper", "UnusedParameter.Local", Justification = SuppressReasons.UsedViaReflection)]
+    private void OnDeserializedMethod(StreamingContext context)
+    {
+        this.ToggleKey ??= new KeybindList();
+    }
 }