+++ TimeSpeed/Framework/AutoFreezeReason.cs
@@ -1,5 +1,3 @@
-using StardewValley;
-
 namespace TimeSpeed.Framework;
 
 /// <summary>The reasons for automated time freezes.</summary>
@@ -8,9 +6,12 @@ internal enum AutoFreezeReason
     /// <summary>No freeze currently applies.</summary>
     None,
 
-    /// <summary>Time was automatically frozen based on the location per <see cref="ModConfig.ShouldFreeze(GameLocation)"/>.</summary>
+    /// <summary>Time was automatically frozen based on the location.</summary>
     FrozenForLocation,
 
-    /// <summary>Time was automatically frozen per <see cref="ModConfig.ShouldFreeze(int)"/>.</summary>
-    FrozenAtTime
+    /// <summary>Time was automatically frozen based on the time of day.</summary>
+    FrozenAtTime,
+
+    /// <summary>Time was automatically frozen before the player passed out.</summary>
+    FrozenBeforePassOut
 }