+++ TimeSpeed/Framework/ModFreezeTimeConfig.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.Serialization;
+using cantorsdust.Common;
 using StardewValley;
 using StardewValley.Locations;
 
@@ -14,6 +16,9 @@ internal class ModFreezeTimeConfig
     *********/
     /// <summary>The time at which to freeze time everywhere (or <c>null</c> to disable this). This should be 24-hour military time (e.g. 800 for 8am, 1600 for 8pm, etc).</summary>
     public int? AnywhereAtTime { get; set; }
+
+    /// <summary>Whether to freeze time before the player passes out (i.e. at 1:50am).</summary>
+    public bool PassOut { get; set; } = false;
 
     /// <summary>Whether to freeze time indoors.</summary>
     public bool Indoors { get; set; } = false;
@@ -44,7 +49,7 @@ internal class ModFreezeTimeConfig
     *********/
     /// <summary>Get whether time should be frozen in the given location.</summary>
     /// <param name="location">The location to check.</param>
-    public bool ShouldFreeze(GameLocation location)
+    public bool ShouldFreeze(GameLocation? location)
     {
         if (location == null || this.ExceptLocationNames.Contains(location.Name))
             return false;
@@ -82,9 +87,15 @@ internal class ModFreezeTimeConfig
     /// <summary>The method called after the config file is deserialized.</summary>
     /// <param name="context">The deserialization context.</param>
     [OnDeserialized]
+    [SuppressMessage("ReSharper", "NullCoalescingConditionIsAlwaysNotNullAccordingToAPIContract", Justification = SuppressReasons.ValidatesNullability)]
+    [SuppressMessage("ReSharper", "UnusedMember.Local", Justification = SuppressReasons.UsedViaReflection)]
+    [SuppressMessage("ReSharper", "UnusedParameter.Local", Justification = SuppressReasons.UsedViaReflection)]
     private void OnDeserializedMethod(StreamingContext context)
     {
-        this.ByLocationName = new(this.ByLocationName ?? new(), StringComparer.OrdinalIgnoreCase);
-        this.ExceptLocationNames = new(this.ExceptLocationNames ?? new(), StringComparer.OrdinalIgnoreCase);
+        this.ByLocationName = new HashSet<string>(this.ByLocationName ?? [], StringComparer.OrdinalIgnoreCase);
+        this.ByLocationName.Remove(null!);
+
+        this.ExceptLocationNames = new HashSet<string>(this.ExceptLocationNames ?? [], StringComparer.OrdinalIgnoreCase);
+        this.ExceptLocationNames.Remove(null!);
     }
 }