+++ AllProfessions/docs/release-notes.md
@@ -1,6 +1,11 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## 1.5.8
+Released 11 September 2025 for SMAPI 4.1.2 or later. Updated by Pathoschild.
+
+* Internal changes to simplify maintenance.
+
 ## 1.5.7
 Released 04 November 2024 for SMAPI 4.1.2 or later. Updated by Pathoschild.